refactor: 트랜잭션 서비스 추상화로 리팩터링
